+++ src/main/java/pl/asie/computronics/integration/railcraft/item/ItemRelaySensor.java
@@ -13,9 +13,9 @@ import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.World;
 import pl.asie.computronics.Computronics;
+import pl.asie.computronics.integration.railcraft.tile.TileLocomotiveRelay;
 import pl.asie.computronics.oc.manual.IItemWithPrefix;
 import pl.asie.computronics.reference.Config;
-import pl.asie.computronics.integration.railcraft.tile.TileLocomotiveRelay;
 import pl.asie.computronics.util.StringUtil;
 
 import java.util.List;